fix : Path 객체 도출
